+++ app/exceptions/configuration_error.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+# Raised when Fab-manager is misconfigured
+class ConfigurationError < StandardError
+end